Intergrate CrawlingMixin and SearchMixin -> movie Mixin
기존에 중복되던 함수들을 movieMixin으로 묶었다.
