+++ django_app/movie/management/commands/get_dummy_data.py
@@ -5,7 +5,7 @@ import requests
 from bs4 import BeautifulSoup
 from django.core.management.base import BaseCommand
 
-from utils.movie import crawlingMixin
+from utils.movie import movieMixin
 
 
 def get_all_moviename(page_num):
@@ -28,15 +28,15 @@ def get_all_moviename(page_num):
 def init_naver_movie():
     movie_arr = [
         get_all_moviename(page_num=page_num)
-        for page_num in range(1, 3)
+        for page_num in range(39, 40)
         ]
 
     for page_num in range(len(movie_arr)):
         time.sleep(10)
         for cnt, movie in enumerate(movie_arr[page_num]):
             if cnt != 0 and cnt % 10 == 0:
-                time.sleep(10)
-            crawlingMixin.insert_db(movie)
+                time.sleep(1)
+            movieMixin.search_movie(movie[1])
             time.sleep(1)
